major update
reworked the api routes to show properly, changed the folder structure to more correctly show the MVC model and moved the server.js file to the main folder.
